moved some const to constants.js
